+++ CHANGELOG.md
@@ -6,3 +6,4 @@
  - create and include the change log
  - update the readme.md and outline basic usage
  - organise the files
+ - rename 'frame' method to 'keyframe'